moved user and issue models to root, fixed migrations, removed picture from issue model
