+++ package.json
@@ -49,6 +49,7 @@
   },
   "dependencies": {
     "react-router-dom": "^6.26.2",
-    "sequelize": "^6.37.3"
+    "sequelize": "^6.37.3",
+    "sequelize-cli": "^6.6.2"
   }
 }
