+++ server/models/user.js
@@ -1,33 +1,33 @@
-import { DataTypes } from "sequelize";
-import sequelize from "../config/connection.js";
+// import { DataTypes } from "sequelize";
+// import sequelize from "../config/connection.js";
 
-const User = sequelize.define(
-  "User",
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    name: {
-      type: DataTypes.STRING,
-    },
-    username: {
-      type: DataTypes.STRING,
-    },
-    email: {
-      type: DataTypes.STRING,
-    },
-    phone: {
-      type: DataTypes.STRING,
-    },
-    password: {
-      type: DataTypes.STRING,
-    },
-  },
-  {
-    timestamps: true,
-  }
-);
+// const User = sequelize.define(
+//   "User",
+//   {
+//     id: {
+//       type: DataTypes.INTEGER,
+//       autoIncrement: true,
+//       primaryKey: true,
+//     },
+//     name: {
+//       type: DataTypes.STRING,
+//     },
+//     username: {
+//       type: DataTypes.STRING,
+//     },
+//     email: {
+//       type: DataTypes.STRING,
+//     },
+//     phone: {
+//       type: DataTypes.STRING,
+//     },
+//     password: {
+//       type: DataTypes.STRING,
+//     },
+//   },
+//   {
+//     timestamps: true,
+//   }
+// );
 
-export default User;
+// export default User;
